Change the discount from int to double (crash when no discount is inputed)

+++ app/src/main/java/sg/edu/rp/c346/id20026955/billplease/MainActivity.java
@@ -50,7 +50,7 @@ public class MainActivity extends AppCompatActivity {
                 String pax = editPax.getText().toString();
                 int NoOfPax = Integer.parseInt(pax);
                 String discount = editDiscount.getText().toString();
-                int dis = Integer.parseInt(discount);
+                double dis = Double.parseDouble(discount);
                 int selectedID = rgPayment.getCheckedRadioButtonId();
 
                 if (SVS.isChecked() == true && GST.isChecked() == false) {
